Extended-Kalman-Filter: Predict and Update Steps
(1) Modify the F and Q matrices prior to the prediction step
based on the elapsed time between measurements.
(2) Call the update step for either the lidar or radar sensor measurement.
Because the update step for lidar and radar are slightly different,
there are different functions for updating lidar and radar.